+++ SurvivalKit.Tests/Events/Mocks/MockEvent.cs
@@ -1,4 +1,4 @@
-﻿using SurvivalKit.Events.Interfaces;
+﻿using SurvivalKit.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Linq;
