+++ SurvivalKit/Events/Environment/SetBlocksEvent.cs
@@ -1,3 +1,4 @@
+using SurvivalKit.Interfaces;
 using System;
 using System.Collections.Generic;
