+++ SurvivalKit/Interfaces/ICancellableEvent.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SurvivalKit.Interfaces
+{
+	public interface ICancellableEvent : IDispatchableEvent
+	{
+	}
+}
