+++ SurvivalKit/Interfaces/IDispatchableEvent.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace SurvivalKit.Events.Interfaces
+namespace SurvivalKit.Interfaces
 {
 	/// <summary>
 	///	Interface all events should implement if they want to be dispatched by the <see cref="EventAggregator"/>.
