+++ SurvivalKit/Interfaces/IEventAggregator.cs
@@ -1,9 +1,9 @@
-﻿using SurvivalKit.Events.Abstracts;
+﻿using SurvivalKit.Abstracts;
 using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace SurvivalKit.Events.Interfaces
+namespace SurvivalKit.Interfaces
 {
 	public interface IEventAggregator
 	{
